Modified Helper classes
